Extract fetchTranslations to separate file
