+++ packages/template-retail-react-app/app/utils/locale.js
@@ -8,36 +8,6 @@
 import PropTypes from 'prop-types'
 
 /**
- * Dynamically import the translations/messages for a given locale
- * @param {string} locale
- * @returns {Promise<Object>} The messages (compiled in AST format) in the given locale.
- *      If the translation file is not found, return an empty object (so react-intl would fall back to the inline messages)
- */
-export const fetchTranslations = async (locale) => {
-    const targetLocale =
-        typeof window === 'undefined'
-            ? process.env.USE_PSEUDOLOCALE === 'true'
-                ? 'en-XA'
-                : locale
-            : locale
-
-    let module
-    try {
-        module = await import(
-            `@salesforce/retail-react-app/translations/compiled/${targetLocale}.json`
-        )
-    } catch (err) {
-        console.error(err)
-        console.log(
-            'Loading empty messages, so that react-intl would fall back to the inline default messages'
-        )
-        return {}
-    }
-
-    return module.default
-}
-
-/**
  * Determine the target locale by comparing the users' preferred locales with the app's supported locales
  * @param {Object} options
  * @param {function} [options.getUserPreferredLocales] - Identify what set of locales the user prefers
